SEC HTML/URL Text extraction
This utilized the span element within the html which only occurs for text. This cleaned up a lot, but will have to verify if we lost any needed info, and possibly still clean some more from here. This is just one example, will need to be run on a list of the URLs